+++ extractor/src/main/scala/com/digitalasset/extractor/Extractor.scala
@@ -12,23 +12,27 @@ import com.digitalasset.extractor.config.{ExtractorConfig, SnapshotEndSetting}
 import com.digitalasset.extractor.ledger.LedgerReader
 import com.digitalasset.extractor.ledger.types.TransactionTree
 import com.digitalasset.extractor.ledger.types.TransactionTree._
+import com.digitalasset.extractor.ledger.LedgerReader.PackageStore
 import com.digitalasset.extractor.targets.Target
 import com.digitalasset.extractor.writers.Writer
 import com.digitalasset.extractor.writers.Writer.RefreshPackages
+import com.digitalasset.extractor.helpers.TemplateIds
+import com.digitalasset.extractor.helpers.TransactionTreeTrimmer
+import com.digitalasset.extractor.helpers.FutureUtil.toFuture
 import com.digitalasset.grpc.adapter.{ExecutionSequencerFactory, AkkaExecutionSequencerPool}
 import com.digitalasset.ledger.api.v1.ledger_offset.LedgerOffset
-import com.digitalasset.ledger.api.v1.transaction_filter.{TransactionFilter, Filters}
+import com.digitalasset.ledger.api.v1.transaction_filter.{Filters, TransactionFilter}
+import com.digitalasset.ledger.api.{v1 => api}
 import com.digitalasset.ledger.client.LedgerClient
 import com.digitalasset.ledger.client.configuration._
-
 import io.grpc.netty.{NegotiationType, NettyChannelBuilder}
+
 import scala.collection.breakOut
 import scala.concurrent.duration._
 import scala.concurrent.Future
 import scala.util.control.NonFatal
 import scalaz._
 import Scalaz._
-
 import scalaz.syntax.tag._
 
 class Extractor[T <: Target](config: ExtractorConfig, target: T) {
@@ -74,7 +78,12 @@ class Extractor[T <: Target](config: ExtractorConfig, target: T) {
 
       _ = log.trace("Handling packages...")
 
-      _ <- fetchPackages(client, writer)
+      packageStore <- fetchPackages(client, writer)
+      allTemplateIds = TemplateIds.getTemplateIds(packageStore.values.toSet)
+      _ = log.info(s"All available template ids: {}", allTemplateIds)
+
+      requestedTemplateIds <- Future.successful(
+        TemplateIds.intersection(allTemplateIds, config.templateConfigs))
 
       streamUntil: Option[LedgerOffset] = config.to match {
         case SnapshotEndSetting.Head => Some(endOffset)
@@ -85,7 +94,7 @@ class Extractor[T <: Target](config: ExtractorConfig, target: T) {
 
       _ = log.info("Handling transactions...")
 
-      _ <- streamTransactions(client, writer, streamUntil)
+      _ <- streamTransactions(client, writer, streamUntil, requestedTemplateIds)
 
       _ = log.info("Done...")
     } yield ()
@@ -104,23 +113,35 @@ class Extractor[T <: Target](config: ExtractorConfig, target: T) {
     system.terminate().map(_ => killSwitch.shutdown())
   }
 
-  private def fetchPackages(client: LedgerClient, writer: Writer): Future[Unit] = {
+  private def fetchPackages(client: LedgerClient, writer: Writer): Future[PackageStore] = {
     for {
-      packageStore <- LedgerReader.createPackageStore(client)
-      _ <- writer.handlePackages(packageStore.valueOr(error => throw new RuntimeException(error)))
-    } yield ()
+      packageStoreE <- LedgerReader.createPackageStore(client): Future[String \/ PackageStore]
+      packageStore <- toFuture(packageStoreE): Future[PackageStore]
+      _ <- writer.handlePackages(packageStore)
+    } yield packageStore
   }
 
-  private def selectTransactions: TransactionFilter = {
+  private def selectTransactions(parties: ExtractorConfig.Parties): TransactionFilter = {
+    // Template filtration is not supported on GetTransactionTrees RPC
+    // we will have to filter out template on client-side.
     val templateSelection = Filters.defaultInstance
-    TransactionFilter(config.parties.toList.map(_ -> templateSelection)(breakOut))
+    val result = TransactionFilter(config.parties.toList.map(_ -> templateSelection)(breakOut))
+    log.info(s"Setting transaction filter: {}", result)
+    result
   }
 
   private def streamTransactions(
       client: LedgerClient,
       writer: Writer,
-      streamUntil: Option[LedgerOffset]
+      streamUntil: Option[LedgerOffset],
+      requestedTemplateIds: Set[api.value.Identifier]
   ): Future[Unit] = {
+    log.info("Requested template IDs: {}", requestedTemplateIds)
+
+    val trim: api.transaction.TransactionTree => api.transaction.TransactionTree =
+      if (requestedTemplateIds.isEmpty) identity
+      else TransactionTreeTrimmer.trim(requestedTemplateIds)
+
     RestartSource
       .onFailuresWithBackoff(
         minBackoff = 3.seconds,
@@ -132,16 +153,14 @@ class Extractor[T <: Target](config: ExtractorConfig, target: T) {
           .getTransactionTrees(
             LedgerOffset(startOffSet),
             streamUntil,
-            selectTransactions,
+            selectTransactions(config.parties),
             verbose = true
           )
           .via(killSwitch.flow)
-          .map(
-            _.convert.fold(
-              e => throw DataIntegrityError(e),
-              identity
-            )
-          )
+          .map(trim)
+          .collect {
+            case t if nonEmpty(t) => convertTransactionTree(t)
+          }
           .mapAsync(parallelism = 1) { t =>
             writer
               .handleTransaction(t)
@@ -157,6 +176,11 @@ class Extractor[T <: Target](config: ExtractorConfig, target: T) {
       .runWith(Sink.ignore)
       .void
   }
+
+  private def nonEmpty(t: api.transaction.TransactionTree): Boolean = t.eventsById.nonEmpty
+
+  private def convertTransactionTree(t: api.transaction.TransactionTree): TransactionTree =
+    t.convert.fold(e => throw DataIntegrityError(e), identity)
 
   /**
     * We encountered a transaction that reference a previously not witnessed type.