+++ extractor/src/main/scala/com/digitalasset/extractor/config/ExtractorConfig.scala
@@ -27,6 +27,7 @@ final case class ExtractorConfig(
     from: LedgerOffset,
     to: SnapshotEndSetting,
     parties: ExtractorConfig.Parties,
+    templateConfigs: Set[TemplateConfig],
     tlsConfig: TlsConfiguration,
     appId: String = s"Extractor-${UUID.randomUUID().toString}"
 ) {
@@ -36,4 +37,11 @@ final case class ExtractorConfig(
 
 object ExtractorConfig {
   type Parties = OneAnd[List, Party]
+}
+
+final case class TemplateConfig(moduleName: String, entityName: String)
+
+object TemplateConfig {
+  implicit val templateConfigOrdering: Ordering[TemplateConfig] =
+    Ordering.by(TemplateConfig.unapply)
 }