+++ extractor/src/main/scala/com/digitalasset/extractor/helpers/FutureUtil.scala
@@ -1,0 +1,28 @@
+// Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.digitalasset.extractor.helpers
+
+import scalaz.{NonEmptyList, Show, ValidationNel, \/}
+import scala.concurrent.Future
+
+object FutureUtil {
+  def toFuture[E: Show, A](a: E \/ A): Future[A] = {
+    val E = implicitly[Show[E]]
+    a.fold(
+      e => Future.failed(new RuntimeException(E.shows(e))),
+      a => Future.successful(a)
+    )
+  }
+
+  def toFuture[E: Show, A](a: ValidationNel[E, A]): Future[A] =
+    a.fold(
+      es => Future.failed(new RuntimeException(formatErrors(es))),
+      a => Future.successful(a)
+    )
+
+  private def formatErrors[E: Show](es: NonEmptyList[E]): String = {
+    val E = implicitly[Show[E]]
+    es.map(e => E.shows(e)).list.toList.mkString(", ")
+  }
+}